Consolidate tile requests from different services
